vhost: insert transport option in the vhost library
This patch inserts a transport option in the vhost library. This way,
the vhost library will be able make use of the extended DPDK's
librte_vhost API, which supports the AF_UNIX and virtio-vhost-user
transports.

The transport choice is being passed as a flag to
`rte_vhost_driver_register()`. The chosen transport is being detected at
runtime based on the provided vhost controller name. The vhost
controller name can be supplied either through the JSON RPC API or
through the static configuration file API. A PCI DomBDF address
(belonging to a virtio-vhost-user device) implies choosing the
virtio-vhost-user transport. In any other case, the AF_UNIX transport is
being used. There is 1:1 correspondence among virtio-vhost-user devices
and vhost controllers, hence we can map a specific virtio-vhost-user
device to only one vhost controller.

SPDK's internal `rte_vhost` copy will not be updated to support the
virtio-vhost-user transport. Therefore, the internal
`rte_vhost_driver_register()` just returns with -1 in case the
virtio-vhost-user transport is requested.

Example usage of the JSON RPC API:

$ scripts/rpc.py construct_vhost_scsi_controller --cpumask 0x1 vhost.0
$ scripts/rpc.py construct_vhost_scsi_controller --cpumask 0x1 0000:00:07.0

+++ lib/vhost/vhost.c
@@ -691,6 +691,8 @@ spdk_vhost_dev_register(struct spdk_vhost_dev *vdev, const char *name, const cha
 	char path[PATH_MAX];
 	struct stat file_stat;
 	struct spdk_cpuset *cpumask;
+	struct spdk_pci_addr pci_addr;
+	uint64_t transport = 0;
 	int rc;
 
 	assert(vdev);
@@ -718,36 +720,51 @@ spdk_vhost_dev_register(struct spdk_vhost_dev *vdev, const char *name, const cha
 		goto out;
 	}
 
-	if (snprintf(path, sizeof(path), "%s%s", dev_dirname, name) >= (int)sizeof(path)) {
-		SPDK_ERRLOG("Resulting socket path for controller %s is too long: %s%s\n", name, dev_dirname,
-			    name);
-		rc = -EINVAL;
-		goto out;
+	if (spdk_pci_addr_parse(&pci_addr, name) == 0) {
+		transport = RTE_VHOST_USER_VIRTIO_TRANSPORT;
 	}
 
-	/* Register vhost driver to handle vhost messages. */
-	if (stat(path, &file_stat) != -1) {
-		if (!S_ISSOCK(file_stat.st_mode)) {
-			SPDK_ERRLOG("Cannot create a domain socket at path \"%s\": "
-				    "The file already exists and is not a socket.\n",
-				    path);
-			rc = -EIO;
+	if (transport != RTE_VHOST_USER_VIRTIO_TRANSPORT) {
+		if (snprintf(path, sizeof(path), "%s%s", dev_dirname, name) >= (int)sizeof(path)) {
+			SPDK_ERRLOG("Resulting socket path for controller %s is too long: %s%s\n", name, dev_dirname,
+				    name);
+			rc = -EINVAL;
 			goto out;
-		} else if (unlink(path) != 0) {
-			SPDK_ERRLOG("Cannot create a domain socket at path \"%s\": "
-				    "The socket already exists and failed to unlink.\n",
-				    path);
-			rc = -EIO;
+		}
+
+		/* Register vhost driver to handle vhost messages. */
+		if (stat(path, &file_stat) != -1) {
+			if (!S_ISSOCK(file_stat.st_mode)) {
+				SPDK_ERRLOG("Cannot create a domain socket at path \"%s\": "
+					    "The file already exists and is not a socket.\n",
+					    path);
+				rc = -EIO;
+				goto out;
+			} else if (unlink(path) != 0) {
+				SPDK_ERRLOG("Cannot create a domain socket at path \"%s\": "
+					    "The socket already exists and failed to unlink.\n",
+					    path);
+				rc = -EIO;
+				goto out;
+			}
+		}
+	} else {
+		if (snprintf(path, sizeof(path), "%s", name) != (int)strlen(name)) {
+			SPDK_ERRLOG("Failed to copy PCI address '%s' for controller.\n", name);
+			rc = -EINVAL;
 			goto out;
 		}
 	}
 
-	if (rte_vhost_driver_register(path, 0) != 0) {
+	if (rte_vhost_driver_register(path, transport) != 0) {
 		SPDK_ERRLOG("Could not register controller %s with vhost library\n", name);
-		SPDK_ERRLOG("Check if domain socket %s already exists\n", path);
+		if (transport != RTE_VHOST_USER_VIRTIO_TRANSPORT) {
+			SPDK_ERRLOG("Check if domain socket %s already exists\n", path);
+		}
 		rc = -EIO;
 		goto out;
 	}
+
 	if (rte_vhost_driver_set_features(path, backend->virtio_features) ||
 	    rte_vhost_driver_disable_features(path, backend->disabled_features)) {
 		SPDK_ERRLOG("Couldn't set vhost features for controller %s\n", name);
